+++ indexer/database/database_transaction.go
@@ -9,6 +9,11 @@ type DbTransaction interface {
 	// RollbackTransaction rollbacks the current transaction instance, returns error if any problem happened using rollback progress
 	RollbackTransaction() error
 
+	// Chain info
+
+	// SetLatestIndexedBlock updates the latest indexed block height of the chain info record with the given chain ID.
+	SetLatestIndexedBlock(chainId string, height int64) error
+
 	// Account
 
 	// InsertOrUpdateRecordsAccount inserts or updates the given accounts into the database.
@@ -18,7 +23,22 @@ type DbTransaction interface {
 
 	// Transaction
 
+	// InsertRecordTransactionsIfNotExists inserts the given transactions into the database.
+	// If the transaction is exists, it will do nothing.
 	InsertRecordTransactionsIfNotExists(txs dbtypes.RecordsTransaction) error
+
+	// InsertRecordsRecentAccountTransactionIfNotExists inserts the given recent account transactions into the database.
+	// If the transaction is exists, it will do nothing.
+	InsertRecordsRecentAccountTransactionIfNotExists(txs dbtypes.RecordsRecentAccountTransaction) error
+
+	// InsertRecordsRefAccountToRecentTxIfNotExists inserts the given references of account to recent transaction into the database.
+	// If the reference is exists, it will do nothing.
+	// The record connects the recent account transaction with the account that has the transaction.
+	InsertRecordsRefAccountToRecentTxIfNotExists(refs dbtypes.RecordsRefAccountToRecentTx) error
+
+	// CleanupZeroRefCountRecentAccountTransaction call procedures to clean-up `recent_account_transaction` records
+	// which have zero referent (`ref_account_to_recent_tx`).
+	CleanupZeroRefCountRecentAccountTransaction() error
 
 	// Failed blocks