+++ indexer/integration_test_util/db_suite_read_record.go
@@ -145,6 +145,137 @@ FROM "transaction" WHERE hash = $1 AND height = $2 AND chain_id = $3
 	return res
 }
 
+// ReadRecentAccountTransactionRecord reads a specific transaction record from `recent_account_transaction` table in database.
+//
+//goland:noinspection SqlNoDataSourceInspection,SqlDialectInspection,SpellCheckingInspection
+func (suite *DatabaseIntegrationTestSuite) ReadRecentAccountTransactionRecord(hash string, height int64, chainId string, optionalTx *sql.Tx) itutildbtypes.RecentAccountTransactionRecord {
+	statement := `
+SELECT
+    chain_id, -- 1
+	height, -- 2
+	hash, -- 3
+	ref_count, -- 4
+	epoch, -- 5
+	message_types -- 6
+FROM recent_account_transaction WHERE hash = $1 AND height = $2 AND chain_id = $3
+`
+
+	var rows *sql.Rows
+	var err error
+	if optionalTx == nil {
+		rows, err = suite.Database.Query(statement, hash, height, chainId)
+	} else {
+		rows, err = optionalTx.Query(statement, hash, height, chainId)
+	}
+
+	suite.Require().NoError(err, "failed to query recent account transaction")
+	defer func() {
+		_ = rows.Close()
+	}()
+	suite.Require().Truef(rows.Next(), "no record found for recent account transaction %s at %d", hash, height)
+
+	var res itutildbtypes.RecentAccountTransactionRecord
+
+	err = rows.Scan(
+		&res.ChainId,                // 1
+		&res.Height,                 // 2
+		&res.Hash,                   // 3
+		&res.RefCount,               // 4
+		&res.Epoch,                  // 5
+		pq.Array(&res.MessageTypes), // 6
+	)
+	suite.Require().NoError(err, "failed to scan recent account transaction")
+	suite.Require().Falsef(rows.Next(), "more than one record found for recent account transaction %s at %d", hash, height)
+
+	return res
+}
+
+// ReadReducedRefCountRecentAccountTransactionRecord reads a specific transaction record from `reduced_ref_count_recent_account_transaction` table in database.
+//
+//goland:noinspection SqlNoDataSourceInspection,SqlDialectInspection,SpellCheckingInspection
+func (suite *DatabaseIntegrationTestSuite) ReadReducedRefCountRecentAccountTransactionRecord(hash string, height int64, chainId string, optionalTx *sql.Tx) itutildbtypes.ReducedRefCountRecentAccountTransactionRecord {
+	statement := `
+SELECT
+    chain_id, -- 1
+	height, -- 2
+	hash, -- 3
+FROM recent_account_transaction WHERE hash = $1 AND height = $2 AND chain_id = $3
+`
+
+	var rows *sql.Rows
+	var err error
+	if optionalTx == nil {
+		rows, err = suite.Database.Query(statement, hash, height, chainId)
+	} else {
+		rows, err = optionalTx.Query(statement, hash, height, chainId)
+	}
+
+	suite.Require().NoError(err, "failed to query recent account transaction")
+	defer func() {
+		_ = rows.Close()
+	}()
+	suite.Require().Truef(rows.Next(), "no record found for recent account transaction %s at %d", hash, height)
+
+	var res itutildbtypes.ReducedRefCountRecentAccountTransactionRecord
+
+	err = rows.Scan(
+		&res.ChainId, // 1
+		&res.Height,  // 2
+		&res.Hash,    // 3
+	)
+	suite.Require().NoError(err, "failed to scan recent account transaction")
+	suite.Require().Falsef(rows.Next(), "more than one record found for recent account transaction %s at %d", hash, height)
+
+	return res
+}
+
+// ReadRefAccountToRecentTxRecord reads a specific ref account to recent tx record from `ref_account_to_recent_tx` table in database.
+//
+//goland:noinspection SqlNoDataSourceInspection,SqlDialectInspection,SpellCheckingInspection
+func (suite *DatabaseIntegrationTestSuite) ReadRefAccountToRecentTxRecord(bech32Address, hash string, height int64, chainId string, optionalTx *sql.Tx) itutildbtypes.RefAccountToRecentTxRecord {
+	statement := `
+SELECT
+    chain_id, -- 1
+	bech32_address, -- 2
+	height, -- 3
+	hash, -- 4
+	signer, -- 5
+	erc20, -- 6
+	nft -- 7
+FROM ref_account_to_recent_tx WHERE bech32_address = $1 AND hash = $2 AND height = $3 AND chain_id = $4
+`
+
+	var rows *sql.Rows
+	var err error
+	if optionalTx == nil {
+		rows, err = suite.Database.Query(statement, bech32Address, hash, height, chainId)
+	} else {
+		rows, err = optionalTx.Query(statement, bech32Address, hash, height, chainId)
+	}
+
+	suite.Require().NoError(err, "failed to query ref account to recent tx")
+	defer func() {
+		_ = rows.Close()
+	}()
+	suite.Require().Truef(rows.Next(), "no record found for ref account to recent tx %s at %d for account %s", hash, height, bech32Address)
+
+	var res itutildbtypes.RefAccountToRecentTxRecord
+
+	err = rows.Scan(
+		&res.ChainId,       // 1
+		&res.Bech32Address, // 2
+		&res.Height,        // 3
+		&res.Hash,          // 4
+		&res.Signer,        // 5
+		&res.Erc20,         // 6
+		&res.NFT,           // 7
+	)
+	suite.Require().NoError(err, "failed to scan ref account to recent tx")
+	suite.Require().Falsef(rows.Next(), "more than one record found for ref account to recent tx %s at %d for account", hash, height, bech32Address)
+
+	return res
+}
+
 // ReadFailedBlockRecord reads a specific failed block record from `failed_block` table in database.
 //
 //goland:noinspection SqlNoDataSourceInspection,SqlDialectInspection,SpellCheckingInspection